+++ lib/gui/app/components/drive-selector/controllers/drive-selector.js
@@ -23,6 +23,7 @@ const constraints = require('../../../../../shared/drive-constraints')
 const analytics = require('../../../modules/analytics')
 const availableDrives = require('../../../../../shared/models/available-drives')
 const selectionState = require('../../../../../shared/models/selection-state')
+const utils = require('../../../../../shared/utils')
 
 module.exports = function (
   $q,
@@ -173,66 +174,17 @@ module.exports = function (
   }
 
   /**
-   * @summary Memoize ImmutableJS list reference
+   * @summary Memoized getDrives function
    * @function
-   * @private
+   * @public
    *
-   * @description
-   * This workaround is needed to avoid AngularJS from getting
-   * caught in an infinite digest loop when using `ngRepeat`
-   * over a function that returns a mutable version of an
-   * ImmutableJS object.
-   *
-   * The problem is that every time you call `myImmutableObject.toJS()`
-   * you will get a new object, whose reference is different from
-   * the one you previously got, even if the data is exactly the same.
-   *
-   * @param {Function} func - function that returns an ImmutableJS list
-   * @returns {Function} memoized function
+   * @returns {Array<Object>} - memoized list of drives
    *
    * @example
-   * const getList = () => {
-   *   return Store.getState().toJS().myList;
-   * };
-   *
-   * const memoizedFunction = memoizeImmutableListReference(getList);
+   * const drives = DriveSelectorController.getDrives()
+   * // Do something with drives
    */
-  this.memoizeImmutableListReference = (func) => {
-    let previousTuples = []
-
-    return (...restArgs) => {
-      let areArgsInTuple = false
-      let state = Reflect.apply(func, this, restArgs)
-
-      previousTuples = _.map(previousTuples, ([ oldArgs, oldState ]) => {
-        if (angular.equals(oldArgs, restArgs)) {
-          areArgsInTuple = true
-
-          if (angular.equals(state, oldState)) {
-            // Use the previously memoized state for this argument
-            state = oldState
-          }
-
-          // Update the tuple state
-          return [ oldArgs, state ]
-        }
-
-        // Return the tuple unchanged
-        return [ oldArgs, oldState ]
-      })
-
-      // Add the state associated with these args to be memoized
-      if (!areArgsInTuple) {
-        previousTuples.push([ restArgs, state ])
-      }
-
-      return state
-    }
-  }
-
-  this.getDrives = this.memoizeImmutableListReference(() => {
-    return this.drives.getDrives()
-  })
+  this.getDrives = utils.memoize(this.drives.getDrives, angular.equals)
 
   /**
    * @summary Get a drive's compatibility status object(s)
@@ -253,9 +205,9 @@ module.exports = function (
    *   // do something
    * }
    */
-  this.getDriveStatuses = this.memoizeImmutableListReference((drive) => {
+  this.getDriveStatuses = utils.memoize((drive) => {
     return this.constraints.getDriveImageCompatibilityStatuses(drive, this.state.getImage())
-  })
+  }, angular.equals)
 
   /**
    * @summary Keyboard event drive toggling
