+++ lib/shared/utils.js
@@ -78,3 +78,62 @@ exports.percentageToFloat = (percentage) => {
 
   return percentage / exports.PERCENTAGE_MAXIMUM
 }
+
+/**
+ * @summary Memoize a function
+ * @function
+ * @private
+ *
+ * @description
+ * This workaround is needed to avoid AngularJS from getting
+ * caught in an infinite digest loop when using `ngRepeat`
+ * over a function that returns a mutable version of an
+ * ImmutableJS object.
+ *
+ * The problem is that every time you call `myImmutableObject.toJS()`
+ * you will get a new object, whose reference is different from
+ * the one you previously got, even if the data is exactly the same.
+ *
+ * @param {Function} func - function that returns an ImmutableJS list
+ * @param {Function} comparer - function to compare old and new args and state
+ * @returns {Function} memoized function
+ *
+ * @example
+ * const getList = () => {
+ *   return Store.getState().toJS().myList;
+ * };
+ *
+ * const memoizedFunction = memoize(getList, angular.equals);
+ */
+exports.memoize = (func, comparer) => {
+  let previousTuples = []
+
+  return (...restArgs) => {
+    let areArgsInTuple = false
+    let state = Reflect.apply(func, this, restArgs)
+
+    previousTuples = _.map(previousTuples, ([ oldArgs, oldState ]) => {
+      if (comparer(oldArgs, restArgs)) {
+        areArgsInTuple = true
+
+        if (comparer(state, oldState)) {
+          // Use the previously memoized state for this argument
+          state = oldState
+        }
+
+        // Update the tuple state
+        return [ oldArgs, state ]
+      }
+
+      // Return the tuple unchanged
+      return [ oldArgs, oldState ]
+    })
+
+    // Add the state associated with these args to be memoized
+    if (!areArgsInTuple) {
+      previousTuples.push([ restArgs, state ])
+    }
+
+    return state
+  }
+}
