+++ tests/gui/components/drive-selector.spec.js
@@ -20,6 +20,7 @@ const _ = require('lodash')
 const m = require('mochainon')
 const angular = require('angular')
 require('angular-mocks')
+const utils = require('../../../lib/shared/utils')
 
 describe('Browser: DriveSelector', function () {
   beforeEach(angular.mock.module(
@@ -27,64 +28,9 @@ describe('Browser: DriveSelector', function () {
   ))
 
   describe('DriveSelectorController', function () {
-    let $controller
-    let $rootScope
-    let $q
-    let $uibModalInstance
-    let WarningModalService
-
-    let controller
-
-    beforeEach(angular.mock.inject(function (
-      _$controller_,
-      _$rootScope_,
-      _$q_,
-      _WarningModalService_
-    ) {
-      $controller = _$controller_
-      $rootScope = _$rootScope_
-      $q = _$q_
-      $uibModalInstance = {}
-      WarningModalService = _WarningModalService_
-    }))
-
-    beforeEach(() => {
-      controller = $controller('DriveSelectorController', {
-        $scope: $rootScope.$new(),
-        $q,
-        $uibModalInstance,
-        WarningModalService
-      })
-    })
-
-    describe('.memoizeImmutableListReference()', function () {
-      it('constant true should return memoized true', function () {
-        const memoizedConstTrue = controller.memoizeImmutableListReference(_.constant(true))
-        m.chai.expect(memoizedConstTrue()).to.be.true
-      })
-
-      it('should reflect state changes', function () {
-        let stateA = false
-        const memoizedStateA = controller.memoizeImmutableListReference(() => {
-          return stateA
-        })
-
-        m.chai.expect(memoizedStateA()).to.be.false
-
-        stateA = true
-
-        m.chai.expect(memoizedStateA()).to.be.true
-      })
-
-      it('should reflect different arguments', function () {
-        const memoizedParameter = controller.memoizeImmutableListReference(_.identity)
-
-        m.chai.expect(memoizedParameter(false)).to.be.false
-        m.chai.expect(memoizedParameter(true)).to.be.true
-      })
-
+    describe('.memoize()', function () {
       it('should handle equal angular objects with different hashes', function () {
-        const memoizedParameter = controller.memoizeImmutableListReference(_.identity)
+        const memoizedParameter = utils.memoize(_.identity, angular.equals)
         const angularObjectA = {
           $$hashKey: 1,
           keyA: true
